pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sprig"
version = "0.1.0"
description = "Teller.io transaction data collection tool"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests[security]>=2.31.0",
    "pydantic>=2.0.0",
    "pydantic-ai>=0.1.0",
    "anthropic>=0.34.0",
    "PyYAML>=6.0.0",
    "flask>=2.3.0",
    "tenacity>=8.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.14.0",
    "pyinstaller>=6.0.0",
]

[project.scripts]
sprig = "sprig.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["sprig*"]

[tool.ruff]
line-length = 100
